+++ lib/item.rb
@@ -1,39 +1,10 @@
+require_relative 'element'
+
+# This class defines items
 class Item
-  attr_accessor :attributes
+  include Element
 
   def initialize(attributes)
     @attributes = attributes
-  end
-
-  def id
-    @attributes[:id].to_i
-  end
-
-  def name
-    @attributes[:name]
-  end
-
-  def description
-    @attributes[:description]
-  end
-
-  def unit_price
-    @attributes[:unit_price]
-  end
-
-  def created_at
-    @attributes[:created_at]
-  end
-
-  def updated_at
-    @attributes[:updated_at]
-  end
-
-  def merchant_id
-    @attributes[:merchant_id].to_i
-  end
-
-  def unit_price_to_dollars
-    @attributes[:unit_price].to_f
   end
 end
